doctr/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-multipart==0.0.6
redis==4.6.0
matplotlib 
pillow 
opencv-python-headless 
defusedxml 
tqdm 
anyascii 
scipy 
shapely 
pypdfium2 
h5py 
pyclipper 
rapidfuzz 
langdetect 
huggingface_hub 
mplcursors
pytesseract
rembg
onnxruntime
gdown